Вынес общий функционал в отдельный файл, добавил проверку на запуск клиента.
